[FIX] base/0.0.0: remove `ON DELETE RESTRICT` even when nullable
There may be cases when the _custom_ column pointing to a model is set
as `ON DELETE RESTRICT` while still allowing for null values.
```
360806_afu=> \d scanner_scenario
+---------------+-----------------------------+-----------------+
| Column        | Type                        | Modifiers       |
|---------------+-----------------------------+-----------------|
...
| model_id      | integer                     |                 |
...
...
Foreign-key constraints:
...
    "scanner_scenario_model_id_fkey" FOREIGN KEY (model_id) REFERENCES ir_model(id) ON DELETE RESTRICT
...
```

Before this change the column `model_id` above is filtered out from the
result to process, and consequently the `ON DELETE RESTRICT` is not
modified. Note that anything with `(a.attnotnull OR t.typnotnull)` and
`confdeltype=n` is no-op in the loop. Thus the processing will be
equivalent to the original query.

upg-360806

closes odoo/upgrade#3651

Signed-off-by: Christophe Simonis <chs@odoo.com>