[IMP] util/orm: create a __cloc_exclude__ server action xmlid with cron
When a cron gets created there's a related server action created
automatically. This SA gets counted as custom lines of code because it's
module is set to the non-existent module `__upgrade__`.

Here the `create_cron` method is improved to add a duplicate xmlid for
the server action with __cloc_exclude__ as the owning module to avoid
the CLOC issues.

closes odoo/upgrade#3580

Signed-off-by: Christophe Simonis <chs@odoo.com>