[FIX] l10n_pe/saas~14.3: move the document types to the right module
They should be moved to the right module as dependencies changed and
as they are defined now in l10n_pe instead of l10n_pe_edi, the records
should have been moved from one module to the other when upgrading.

It is important as well we keep the original document type on the invoices
already sent to the government.

We stumbled upon the problem adding an extra document type needed for PE in:
https://github.com/odoo/odoo/pull/91176

Original move:
https://github.com/odoo/odoo/pull/60148
(and related https://github.com/odoo/enterprise/pull/15213)

closes odoo/upgrade#3664

Signed-off-by: Christophe Simonis <chs@odoo.com>