[IMP] base/saas~15.5: remove sale_project_enterprise
This module was created to contained a customization done in JS to
support mobile view for Sales Section displayed in the right side panel
shown in the project update kanban and list view. Since this
customization is no longer needed from the convertion of the views in
OWL, this module is empty and so it is removed.

task-2713538

Related PRs:

* Community: odoo/odoo#98380
* Enterprise: odoo/enterprise#30565

closes odoo/upgrade#3787

Signed-off-by: Christophe Simonis <chs@odoo.com>