[IMP] mail,sms/saas~15.5: add `template_fs` in existing templates
With linked task, we have added a feature to restore the (mail/sms) template.
To do so, we save the file path within the record itself when a module is
being installed (stored within `template_fs` field). But for the existing
records, we do it manually with this commit.

task- 2231977
related PR- https://github.com/odoo/odoo/pull/83759

closes odoo/upgrade#3674

Related: odoo/odoo#83759
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>