[IMP] purchase_stock/saas~15.5: remove price different account
The community PR is mainly to automatically correct stock value and
accounting when there is a price difference between a purchase.order
and the related invoice. Price difference account was use to
partially fix this situtation.

Price different account is not used anymore since the correction are
automaticaly made with the layers. So remove the associated view.

Task-2081607
See odoo/odoo/pull/99411

closes odoo/upgrade#3878

Related: odoo/odoo#99411
Signed-off-by: Christophe Simonis <chs@odoo.com>