[IMP] base/14.0.1.3: avoid unnecessary calls to fixup_m2m
`update_db()` is called many times per model in each upgrade. In the
13->14 migration, this translates into multiple calls to `fixup_m2m()`
on every m2m table.

As per #1940, the aim of this is to enforce restrictions that were not
enforced prior to 14.0. This does not need to be done more than once
per relation.

Tested on upg-244576, there are 1614 calls to `fixup_m2m()` coming from
here, on 199 unique tables. This accounts for 9 minutes out of a 55
minutes upgrade, down to 199 calls in 1 minute after this patch.

closes odoo/upgrade#3699

Signed-off-by: Christophe Simonis <chs@odoo.com>