[IMP] payment/saas~15.5: change of field's name and deletion of field
Changing the field name "country_ids" to "available_country_ids" to
be more on accord with its purpose.

Removing the currency_id field from the sepa_direct_debit acquirer,
as we are adding a field with the same definition to the
payment.acquirer model, this field is no longer required.

See the community task for more details.

See - odoo/odoo#82411
See - odoo/enterprise#24412

closes odoo/upgrade#3703

Signed-off-by: Christophe Simonis <chs@odoo.com>