[FIX] stock/saas~14.4: revert description_pickingin/out fields as text
because this was changed in Task-2499504.
The fields `description_picking` and `description_pickingout` are
from the `product.template` model and fields type is `text` in
version 14.0 and the same field type is `text` in version 15.0.

The field type from `html` to `text` has been changed in
version 15.0: odoo/odoo@50addee

OPW- 2919847

closes odoo/upgrade#3706

Signed-off-by: Christophe Simonis <chs@odoo.com>