[FIX] base/tests: ignore search views in the crawler
Odoo tries to find/get a default view when `search` is added to the
`view_mode` list of windows actions. The crawler on the other hand
outputs an error to the logs.

```
Traceback (most recent call last):
  File "/tmp/tmplkpmzm_k/migrations/base/tests/test_mock_crawl.py", line 193, in crawl_menu
    self.mock_action(action_vals)
  File "/tmp/tmplkpmzm_k/migrations/base/tests/test_mock_crawl.py", line 320, in mock_action
    mock_method(model, view, fields_list, domain, group_by)
TypeError: mock_view_search() takes 4 positional arguments but 6 were given
```

This error can be reproduced easily:
* Install contacts app in 14.0
* Edit the window action for Contacts app, add `search` to the list
* Upgrade to 15.0
It does not fail, but the error appears in the logs.
(example test upgrade 1246)

Although this error does not cause an upgrade failure, because the
action will fail since the beginning, it is still an error we can avoid.
The search views are anyways mocked by the crawler.
https://github.com/odoo/upgrade/blob/9292c89e39c7747345f1c8abe3772578d4fc0d06/migrations/base/tests/test_mock_crawl.py#L275

upg-364919

closes odoo/upgrade#3716

Signed-off-by: Christophe Simonis <chs@odoo.com>