[FIX] base/0.0.0: fix restore standard views after children
When a standard view fails it is not enough to restore it only when
failing with all children disabled. We need to check, and restore, it
also after enabling its children.

For example if we remove the field `l10n_mx_edi_colony_code` from the
view `l10n_mx_edi.mx_partner_address_form`
https://github.com/odoo/enterprise/blob/bcc74dad6c7ee4db666d4a99b4bf9ce032c6a512/l10n_mx_edi/data/res_country_data.xml#L774
and set it to `noupdate=True`. Then in v14.0 the view
`l10n_mx_edi_extended.mx_partner_address_form_inherit_l10n_mx_edi_extended`
fails
https://github.com/odoo/enterprise/blob/9efba097ac41428a44307cc706bb6865f2834ecb/l10n_mx_edi_extended/views/res_partner_views.xml#L10-L18

In this case it's not enough to restore the child view in 14.0 because
there was no such view in 13.0. The right solution is to restore the
parent view, but that view is not in error _with its children disabled_.
An extra check with the children re-enabled shows that the view is still
in error and a restore from FS does the job.

upg-363790

closes odoo/upgrade#3708

Signed-off-by: Christophe Simonis <chs@odoo.com>