[FIX] account/saas~14.3.1.1: fix output of all hints
If no single country can be guessed, the user is presented a list with clues
found so far to aid them in setting the country manually. The company's ID
was missing from that output and is now added. The user is shown the clue
message only if clues are actually available.
Additionally, the clue reason is printed before the countries, and the
countries are being shown by name, to increase readability.

Before:
```
Clues found so far are ([company ID: (country ID, reason for clue)]): [({134, 1, 138, 13}, "the company's currency")]
```

Now:
```
Clues found so far are ([company ID: (reason for clue, [countries])]): [1: (the company's currency, ['Luxembourg',, 'Andorra, Principality of', 'Monaco', 'Austria'])]
```

closes odoo/upgrade#3753

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>