[IMP] base/saas~15.5: fix trigram indexes
We added trigram index for char fields since
https://github.com/odoo/odoo/pull/83015. But if `unaccent` is installed
in the database (and isn't force to `False` on field),
these new trigram indexes is pointless and cost a
lot for nothing (almost nothing, it still can be used for equality
operator but in this case a btree will be far more efficient).

Now trigram indexes are created with unaccent, but
we should remove unusable ones (without unaccent)
from the database and let the ORM recreate correctly the new indexes.

odoo/odoo#95943
task-2551518

closes odoo/upgrade#3736

Signed-off-by: Christophe Simonis <chs@odoo.com>