[FIX] base/saas~11.5: fix ambiguous query
For optimization purposes, 1ca4ae5b3d56da6e6350243fefe437a7522ee040
changed part of a query from

```
AND id NOT IN
(
    SELECT res_id
      FROM ir_model_data
     WHERE model = %s
       AND COALESCE(module, '') NOT IN ('', '__export__')
)
```

to

```
AND NOT EXISTS
(
    SELECT 1
      FROM ir_model_data
     WHERE model = %s
       AND COALESCE(module, '') NOT IN ('', '__export__')
       AND id=res_id
)
```

The two are not equivalent, as in the second case, `id=res_id` will
be interpreted as `ir_model_data.id = ir_model_data.res_id`. We need to
explicitly define `id` as being that of the table of the main query.

closes odoo/upgrade#3751

Signed-off-by: Christophe Simonis <chs@odoo.com>