[FIX] base/0.0.0: create missing xmlids for cron server acts
In 10.saas~14 we made ir.cron inherit ir.actions.server
https://github.com/odoo/upgrade/blob/ed75ead14e4c7042069fbdb93b1990c334e78f15/migrations/util/_inherit.py#L547-L549
see also odoo/odoo@1c79c5ee3050ad8ba907aeda70c0d1f3dbe837a0

We create the associated server actions for existing crons in
https://github.com/odoo/upgrade/blob/24989364f5d8148e783330e5afc35a4e565a18c9/migrations/base/10.saas%7E14.1.3/pre-30-cron.py#L10-L17
but we are missing the xmlids for the actions.

This was not a problem, but it is now because the CLOC will count those
server actions towards the custom code if they do not have an xmlid.

To fix the issue we add a 0.0.0 fix that should ensure this is corrected
for already migrated DBs.

upg-369674
opw-2948606

closes odoo/upgrade#3760

Signed-off-by: Christophe Simonis <chs@odoo.com>