[FIX] website/9.0: set website_id for modified views
In 8.0 there was no `website_id` column. Since 9.0 we support multi
website. Before this patch we were leaving the new `website_id` column
of `ir_ui_view` empty _always_. This has a nasty side effect later on
for views that were manually modified. In saas~14.5 we remove some
xmlids but the check relies on `website_id` column
https://github.com/odoo/upgrade/blob/7cd4c8734811bdacc3e364d74ee9a0ece1b9c0d9/migrations/website/saas%7E14.5.1.0/pre-10-contactus.py#L29-L37
https://github.com/odoo/upgrade/blob/7cd4c8734811bdacc3e364d74ee9a0ece1b9c0d9/migrations/website/saas%7E14.5.1.0/pre-10-contactus.py#L69-L76
thus it fails to mark the contact us page as custom when, in fact, it
is.

This commit aims to fix the issue at its source. In another commit we
will handle already migrated DBs.

upg-370083
tbg-307

closes odoo/upgrade#3786

Signed-off-by: Christophe Simonis <chs@odoo.com>