[IMP] l10n_cl_edi_boletas/saas~15.5.1.0: remove daily sales book model
We send the boletas already through the electronic invoicing
system, but until the 1st of August 2022 it was still required to
send a monthly summary as well.  So, as the legal requirement
disappears, we can just remove the model assembling the boletas per month.

Ideally, we just keep the linked attachments. (with the XMLs)

Linked PR: https://github.com/odoo/enterprise/pull/29559

closes odoo/upgrade#3797

Signed-off-by: Christophe Simonis <chs@odoo.com>