[IMP] sale_management/saas~13.3: updating pricelist sequence
during the migration, new pricelists are created in version
saas-13.3 using sale_order_template.
https://github.com/odoo/upgrade/blob/master/migrations/sale_management/saas~13.3.1.0/pre-migrate.py#L131

At that time, the new pricelist gets a sequence of 1 by default.

So pricelist in production has a higher sequence which is shown in afterward.
When a customer creates the sale order, selecting the pricelist

becomes difficult because of the altered sequence.

Issue link:
opw-2949161

closes odoo/upgrade#3802

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>