[FIX] base/0.0.0: `if_unchanged` take into account upgrade start time
When we check if a record was changed since creation we currently use
`write/create_date` difference. This has the drawback that if we change
a record _during_ the upgrade it may make the difference bigger than 1
minute (current default interval to determine if it was changed). In
this case the record would be considered as _changed_ while in fact for
the purpose of being or not a standard record we should _still_ consider
this as _originally_ un-changed (since we only change a record during
the upgrade if it was originally consieder as unchanged).

To illustrate why this is an issue, the script
https://github.com/odoo/upgrade/blob/975eeb1cbc5868419edcbeab996bb298adb12f8a/migrations/website/12.0.1.0/pre-views.py#L6-L12
marks as noupdate=True records that have not been changed. That works
only if we do not make any change to the record during the upgrade
before the execution of that script.

For example in upg-371614 we install `website` in v 10.0 during the
upgrade, which creates the `contactus` page. We then later (>1 minute)
update the view. From that point on the record will be consideres as
_changed_, while in fact it has nothing custom on it, and it should be
updated as a standard record all the way up to 15.0

This is why some upgrades fail on
website/saas~14.5.1.0/post-define-contactus-details.py
tbg-307

Part-of: odoo/upgrade#3811