[IMP] website_sale/9.0: move m2m fixup to module
The script for m2m fixups in base touches relations from many modules.
The relation `product_public_category_product_template_rel` is seen in many
databases to exist even when `website_sale` is not installed. In this case it
would be broken and have a single fk, while the other fk table would not exist.

closes odoo/upgrade#3816

Signed-off-by: Christophe Simonis <chs@odoo.com>