[ADD] account_invoice_extract/saas~15.5: rename `selected_status`
`selected_status` used to indicate if a box was selected by the OCR or not.

It was set to either:
 - 0 => no selection
 - 1 => OCR selected, no user selection
 - 2 => OCR & user selected

It was unnecessary as there is already a field `user_selected` to know if
the box is currently selected by the user.

The new field `ocr_selected` is boolean that indicates if the box is
selected by the OCR.

Enterprise PR: odoo/enterprise#30931
Related task: #2858695

closes odoo/upgrade#3851

Related: odoo/odoo#99439
Signed-off-by: Christophe Simonis <chs@odoo.com>