[IMP] planning{_calendar}/saas~15.5: remove module,view,action
This commit removes the planning_calendar module as the module is
removed in the associated enterprise PR.

It also ensure that the transition from employee_id to resource_id in
`planning.role` is made smoothly by migrating the existing data.

Lastly, it removes the planning.tag_ids and associated view, action and
menu.

task-2972953

Related PR:
- Enterprise: odoo/enterprise#31172

closes odoo/upgrade#3877

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>
Co-authored-by: damr <damr@odoo.com>