[IMP] account_edi_ubl_cii/16.0: always generate Factur-X silently
Factur-X is now always generated when printing any invoice.

Uncheck the "Factur-X PDF/A-3" option on the journals for everyone
since this would then triggers the validation of the invoice and potentially
display warnings on the move.

Since this option also controls the PDF/A-3 conversion, remove the system
parameter 'ir_config_parameter_use_pdfa'.

task-2957808
https://github.com/odoo/odoo/pull/99699
https://github.com/odoo/enterprise/pull/31851

closes odoo/upgrade#3940

Signed-off-by: Christophe Simonis <chs@odoo.com>