[IMP] sign/saas~15.3: create missing sign_request for shared templates
Before this commit, the shared templates that were shared in 15.2 and older
could not be signed anymore. The controller /sign/<share_link> was removed.
If a partner wanted to use an old link, it would result in a 404 error.

The enterprise commit puts back the controller.
This upgrade modification create the necessary records to allow the use of
the updated controller.

enterprise PR: https://github.com/odoo/enterprise/pull/31130

closes odoo/upgrade#3871

Signed-off-by: Christophe Simonis <chs@odoo.com>