[FIX] base/0.0.0: avoid updating `write_date`
The mechanism in `if_unchanged` relies on checking that the `write_date`
of the record has not been changed after the upgrade. Here we keep the
original `write_date` while forcing the update of `arch_fs`.

The issue is relevant for DBs coming from versions <10.0. In such cases
all views would get a `write_date` way bigger than the `create_date`.
Moreover the `write_date` would be post upgrade start time, which we
interpret as the view being untouched unconditionally. This is wrong if
the view was indeed modified after creating (but still before the
upgrade start time).

For example upg-377043 from 8.0 to 15.0. In the original dump:
```
377043_afu=> select id,write_date,create_date from ir_ui_view where id=291
+------+----------------------------+----------------------------+
| id   | write_date                 | create_date                |
|------+----------------------------+----------------------------|
| 291  | 2019-03-06 10:17:15.375448 | 2015-05-15 14:27:03.274412 |
+------+----------------------------+----------------------------+
```
After upgrade to 9.0:
```
377043_afu_9.0=> select id,write_date,create_date from ir_ui_view where id=291
+------+----------------------------+----------------------------+
| id   | write_date                 | create_date                |
|------+----------------------------+----------------------------|
| 291  | 2019-03-06 10:17:15.375448 | 2015-05-15 14:27:03.274412 |
+------+----------------------------+----------------------------+
```
After upgrade to 10.0 without this patch:
```
377043_afu_9.0_10.0=> select id,write_date,create_date from ir_ui_view where id=291
+------+----------------------------+----------------------------+
| id   | write_date                 | create_date                |
|------+----------------------------+----------------------------|
| 291  | 2022-09-08 11:21:03.974632 | 2015-05-15 14:27:03.274412 |
+------+----------------------------+----------------------------+
```
and with this patch:
```
377043_afu_9.0_10.0=> select id,write_date,create_date from ir_ui_view where id=291
+------+----------------------------+----------------------------+
| id   | write_date                 | create_date                |
|------+----------------------------+----------------------------|
| 291  | 2019-03-06 10:17:15.375448 | 2015-05-15 14:27:03.274412 |
+------+----------------------------+----------------------------+
```

Thus upgrading to 10.0 without this patch would make this view to be
assumed as unchanged (since its `write_date` is bigger than the upgrade
start time) while in fact the original data clearly shows that this view
was modified.

closes odoo/upgrade#3874

Signed-off-by: Christophe Simonis <chs@odoo.com>