[IMP] sale_timesheet/saas~15.3: remove allocated_hours calculation
Before this commit, the `allocated_hours` on `project.project` model
was recomputed when the quantity ordered or the UoM are changed on
the  SOL linked to the project. This behavior is only wanted when
the user confirms the quotation to have a SO, after that, the
allocated_hours should not be recomputed because the project manager
could change the value and loses his changes because a Sales person
has changed the quantity ordered on the SOL linked.

This commit removes the migration script to compute the
`allocated_hours` to assume the value set on the project is the one
expected by the project manager.

Related PR: odoo/odoo#99785

task-2955983

closes odoo/upgrade#4003

Signed-off-by: Christophe Simonis <chs@odoo.com>