[IMP] account_intrastat/saas~15.5: supplementary units on products
See https://github.com/odoo/enterprise/pull/30828

In relation to a commit that adapts the way the supplementary units
relate to products and intrastat fields on products are stored in
general in 'account_intrastat', this commit:

Removes fields on the product category, and removes their related views.

Adapts fields on the product template and product variants. These
adaptions represent the relocation of intrastat fields such that they
exist exclusively on the product variant (rather than existing on the
product variant, product template and product category). Data is moved
from the template fields down to the variant fields when no value is
present in the variant fields.

closes odoo/upgrade#3886

Task-id: 2884512
Related: odoo/enterprise#30828
Signed-off-by: Christophe Simonis <chs@odoo.com>