[FIX] base/9.0: reorder removing and adding deps
The dependencies of module `project_timesheet` are not removed
until after the same module is added as a dependency for `project_issue_sheet`,
which will wrongly mark modules `hr_timesheet`, `sale_contract`, `procurement`
as dependencies for `project_issue_sheet`, and also mark them as `to install` if
they are uninstalled and `project_issue_sheet` is installed.

Also, the module `hr_timesheet_invoice` is removed in v9, but its removal
is delayed in case module `sale_contract` is installed, and then it would
be removed in the pre- script of `sale_contract`.

In the case where `sale_contract` is not installed, and `project_issue_sheet`
is installed: `sale_contract` is marked as `to install`. This scenario will not
remove `hr_timesheet_invoice` in base, and will not run the pre script of
sale_contract either to remove it there.

closes odoo/upgrade#3899

Signed-off-by: Christophe Simonis <chs@odoo.com>