[IMP] appointment_{crm, hr_recruitment}/saas~16.2: move fields
Move the opportunity_id field into the appointment.invite model.
It is now possible to store the lead in an appointment invitation
for all category of appointment type and not only the custom one.
Do the same for the applicant_id field in the appointment_hr_recruitment
module.

task-2984657
ENT PR: odoo/enterprise#32945

closes odoo/upgrade#3974

Related: odoo/odoo#106546
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>