[FIX] base/0.0.0: migrate ir_translation only if it exists
The ir_translation table doesn't exist anymore in 16.0, so the upgrade
from 16 to master will always fail as multiple scripts try to clean up
various bits of ir_translation.

Gate them on whether the table exists.

Some of these scripts may need to be reimplemented on top of the new
translations implementation (rm-unpaired seems like it still makes
sense in the new model).