[FIX] base/0.0.0: altering binary columns increases DB size
When changing the columns type from binary to varchar makes the DB grow
too much. It seems PG keeps the original table (or part of it) in the
transaction. Since we are updating many columns this makes the DB grow
too much.

For upg-382624 the DB grows from <3GB to 1100GB (a terabyte!) before
dying with disk full error.

The solution is to commit the cursor. Which leads to the other issue:
this operation takes a lot of time altering the same table many times
for multiple columns (project_task, >400 columns are being updated).
We can consolidate the queries per table and run them in parallel, which
performs the desired commit.

upg-382624 (others are also affected)

closes odoo/upgrade#3929

Signed-off-by: Christophe Simonis <chs@odoo.com>
Co-authored-by: "Jaimin Shah" <jash@odoo.com>