[FIX] base/0.0.0: unlink excepted field
suppress_upgrade_warnings allow to silent some errors emitted by
upgrade script when removing a field.

Right now the field is not removed meaning that the database will keep
the field and this can lead to warning later if the field is relational
and pointing on a model being removed.

    INFO    Deleting 8452@ir.model.fields (knowledge.field_knowledge_article__cover_image_id)
    INFO    Field unlink knowledge.article.cover_image_id explicitly ignored, skipping
    INFO    Deleting 641@ir.model (knowledge.model_knowledge_cover)
    INFO    Model unlink knowledge.cover explicitly ignored, skipping
    INFO    Field unlink knowledge.article.cover_image_id explicitly ignored, skipping
    INFO    The model knowledge.cover could not be dropped because it did not exist in the registry.
    ...
    WARNING Field knowledge.article.cover_image_id with unknown comodel_name 'knowledge.cover'

With this commit the skipped fields will fallback on the orm behavior
and remove the field from the database.

closes odoo/upgrade#3933

Signed-off-by: Xavier Dollé (xdo) <xdo@odoo.com>