[FIX] analytic/saas~15.5.1.1: check company of analytic account
Before saas~15.5.1.1, analytic lines were allowed to have a different
company than their analytic account.
It should not be allowed and a check constraint was added but was forgotten
for the upgrade script.

This commit nulls the company of the analytic account if it has a
different company than one of the lines associated.

closes odoo/upgrade#3943

Signed-off-by: Nicolas Seinlet (nse) <nse@odoo.com>