[FIX] account/tests: Allow comparison for large numbers
While fetching the values from the query, the exponential value
mismatches with the actual value inspite of values being the
same while evaluating such that '-1.4563600133739037e+19'
mismatches with '-14563600133739037000'. Fix provides
evaluation of prepared value of balance as float and
allows it to compare as scientific notation with the
data coming from invariant using 'float_compare'.

closes odoo/upgrade#3951

Signed-off-by: Christophe Simonis <chs@odoo.com>
Co-authored-by: Alvaro Fuentes <afu@odoo.com>