[FIX] payment/9.saas~11: Rename xmlid of payment module
The module of 'payment_acquirer_sips' record 'payment_sips' was renamed
to 'payment' in Version 9.saas~11.1.0 because of that during upgrade record
'payment_sips.payment_acquirer_sips' would be deleted therefore the not null
constraint on acquirer_id of 'payment_transction' was violated and
it blocks the upgrade process.

For this update the module of related record so it would not be deleted.

task-3000966

closes odoo/upgrade#3963

Signed-off-by: Christophe Simonis <chs@odoo.com>