[ADD] {base,website}/16.0: convert views and fields to Bootstrap 5
Bootstrap has been migrated from version 4.* to version 5.1.3
in Odoo version 16.0.

HTML and XML data in database records need to be converted as well,
especially website views (COWed or custom pages) and HTML fields
that could contain code rendered in the frontend.

The conversion is done leveraging `util.convert_bootstrap` module.

For views, only customized website qweb templates are converted.
That means views created through the copy-on-write (COW) mechanism,
that don't have an associated xmlid, at all or from standard modules.

Data in HTML fields in other models is converted with the help of
`util.snippets` functions.

The migration script runs in the pre- stage of `base` module upgrade,
so that the existing data is converted before any new one is loaded
(attempting to convert from BS4 to BS5 data that's already BS5 code
likely results in broken code).

A test for the BS4 to BS5 conversions is implemented for xml views
and runs for the `website` module:
- the test checks that the view "BS Debug" gets properly converted
  to BS5, checking if it matches the known manually converted one.
- additionally, a view made from chunks of Bootstrap 4 and 5 code
  is built and then converted, and each chunk checked to ensure the
  converted version matches the known manually converted one.

See odoo/odoo#95450
See odoo/enterprise#29195

closes odoo/upgrade#3994

Signed-off-by: Christophe Simonis <chs@odoo.com>
Co-authored-by: Benoit Socias <bso@odoo.com>