[ADD] util.convert_bootstrap: helpers to convert bootstrap versions
New util.convert_bootstrap module with helper functions and classes
that implement converting Bootstrap code in XML/HTML document from
one version to another.

Qweb `t-att(f)-` attributes are also supported, but the conversions
might not be perfect because the expressions are not parsed, and the
changes are done through regexes.

Currently only conversion from Bootstrap version 3.0 to 5.0 have been
implemented, and only conversion from 4.0 to 5.0 have been tested.

The implementation uses a `BootstrapConverter` class that handles
loading and parsing of XML/HTML documents from either files, strings
or already parsed lxml etrees, and holds the conversions definitions
lists for each destination Bootstrap version.
The conversion operates by stepping through these definitions, and
matching elements with the given xpaths, then applying the necessary
operations on them to migrate them.
The operations are instances of subclasses of `ElementOperation`,
these hold the actual business code to manipulate the lxml tree, and
are instantiated in the conversions definition list to provide the
operation parameters for the specific step.

Additional implementation details are provided in docstrings and
comments throughout the code.

Part-of: odoo/upgrade#3994