[IMP] website/saas~16.1: view_hierarchy: QWeb views to client action
Before this commit, `view_hierarchy` extends the QWeb view to see
the hierarchy of views.

With the evolution of the javascript framework, this use case has
now more sense in a client action rather than an extension of a
QWeb view.

This commit refactor `view_hierarchy` into a client action.

See odoo/odoo#103477

closes odoo/upgrade#4016

Signed-off-by: Christophe Simonis <chs@odoo.com>