[FIX] util/pg.py: fix sequence value when a table is empty
When a table is empty, `setval(<sequence>, 0)` will fail since $0 \notin [1,
2^{63}-1]$

```
psycopg2.errors.NumericValueOutOfRange: setval: value 0 is out of bounds for
sequence "l10n_mx_edi_res_locality_id_seq" (1..9223372036854775807)
```

A solution is to set the third parameter of `setval()` to `false` and increment
the value of `COALESCE(max(id), 0)` by 1.

In this case, when a table is empty, the current value of the sequence object
will be `1` and the value of `nextval()` at the next (first) insert will be
also `1`.

upg-416934

closes odoo/upgrade#4036

Signed-off-by: Christophe Simonis <chs@odoo.com>