[IMP] account/saas~13.4.1.1: warn of posting draft entries
Due to the refactoring of reconciliation in 13.4, the upgrade scripts
will post draft journal entries that have reconciled journal items.
Unexpected posting of entries is undesirable from an accounting
perspective, so the purpose of this change is to at least warn the
customer that it will happen. Then they can take action before the
upgrade if needed.

opw-3042677
upg-383251

closes odoo/upgrade#4052

Signed-off-by: Christophe Simonis <chs@odoo.com>