[IMP] sale_subscription/saas~16.1: empty origin_order_id of 1st contract
Before this commit, origin_order_id of the first contract of a sequence is
equal to the id of the contract. It prevent looking for new contract easily
in reports.

This PR make sure that the origin_order_id of the first contract is NULL.

closes odoo/upgrade#4049

Enterprise: https://github.com/odoo/enterprise/pull/33520
Taskid: 3048181
Related: odoo/enterprise#33520
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>