[IMP] account/16.0.1.2: catch edge cases when renaming account codes
When renaming the account code to alphanumeric characters only.
We create a few edge cases.
- `123` and `␣123␣` are both converted to `123`.
- `123-456`, `123,456` and `123␣456` are all converted to `123.456`.

As we can create an infinite number of duplicates by playing with
special characters or white spaces. The idea in this PR is to update
the values that will create duplicates **before** applying the
regex to the whole column.

To do this, a CTE of the duplicates is created
In these specific cases, we will update the code for them
by adding `.dupX` to the end of the code where `X` is a number.
We also notify the customer in the migration reports.

Currently 5 upgrades could benefit from this PR.
upg-418480
upg-418373
upg-417887
upg-417762
upg-417097
see the traceback group

closes odoo/upgrade#4094

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>