[FIX] website/16.0: properly convert SCSS custo made on assets_common
Convert all ir.asset records and ir.attachment records which are the
result of a SCSS customization made via the HTML/SCSS editor. Indeed,
in 16.0, with [1] later fixed by [2], the assets_common bundle stopped
being used for frontend screens as all the common files have been put
inside the assets_frontend bundle directly. Custo that were made to
those files were marked as being part of assets_common and now should be
marked as being part of assets_frontend.

This is especially important for the automatic SCSS custo that are made
while changing some themes configuration (like your website font-size).
Those are made to sub-assets (typically _assets_primary_variables) but
were still marked as part of assets_common and should now be marked as
part of assets_frontend to allow further theme customizations.

Note: this does not handle the possibility of a customization that was
made in a common file that does exist in the frontend bundle anymore.
That is a recurring problem that happens at every migration, a generic
solution would need to be found if not already (when the assets were
regrouped at [1], as a first step for 16.0, all common files were added
in frontend, none was entirely removed from the bundle).

[1]: https://github.com/odoo/odoo/commit/08dcbc8f1def06476d025b110049ee596e906816
[2]: https://github.com/odoo/odoo/commit/15e503a841d757e622c8ef2070043b766b28bfde

closes odoo/upgrade#4063

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>