[FIX] website{,_enterprise}/16.0: remove assets from COW layout
Website layouts could be copied and modified and have references to asset
bundles that should not be called from that layout anymore.
Removing the reference to the bundles, this will keep the customization
and allow the upgrade and the tests to pass.

The COW views would also not have page options, which are needed
in 'website_enterprise'.

closes odoo/upgrade#4067

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>