[FIX] sale_subscription/16.0.1.1: remove analytic tag columns in end step
In 16.0, the account.analytic.tag model has been removed, so
sale.order.template,tag_ids and sale.order,account_tag_ids fields are
removed too.
However, removing them in pre-migrate prevents other modules that depend
on sale_subscription to reorganize the information in the migration.
This commit removes the field in pre-migrate, but keeps the column until
the end step, so that other migration scripts can save this information
differently.