[FIX] account/saas~13.4.1.1: retain memo in payments
From Odoo 14.0 we have move_id linked with the bank
statement lines which will fix the link between account.bank.
statement.line and account.move. As a part of payment
refractoring, Memo('ref') in account.move is updated from
Reference('ref') in account.bank.statement.line. Since in most
cases 'ref' is empty in bank statements, it is updated back
with the same in the moves and the value is lost. We have removed
this updation to prevent data loss

closes odoo/upgrade#4081

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>