[FIX] account_asset/saas~12.3.1.0: Fix rounding issue for imbalance
The mismatch for Upgrade Asset journal in account move
lines arised in case when the decimal places are
greater than the current currency decimal places and the
case gets fulfilled such that the asset and company's
currency are the same. Fix rounds the amounts which gets
added to balance the entries for that journal which user
amends.

closes odoo/upgrade#4089

Signed-off-by: Christophe Simonis <chs@odoo.com>