[FIX] util/convert_bootstrap,website/tests: convert missing form classes
In [1], rules were added to convert Bootstrap classes to BS5. However,
the `form-control-file` and `form-control-range` classes, becoming
`form-control` and `form-range` respectively, were forgotten.

This commit adds their conversion and adapts the tests for them.

The related odoo and enterprise PRs can be found at [2] and [3]. They
remove the `form-control-file` class occurrences from the codebase
(there is no occurrence of the `form-control-range` class).

[1]: https://github.com/odoo/upgrade/commit/66dc8796563aef2e7c294923b96ee971b41cb387
[2]: https://github.com/odoo/odoo/pull/105872
[3]: https://github.com/odoo/enterprise/pull/34495

task-3071151

closes odoo/upgrade#4099

Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>