[IMP] stock/saas~16.1: replace label wizard's move_line_ids
When printing labels in a `ready` pickings, if no quantities are set it
will only print a blank page, since no move_lines could be found.
This change aims to print the requested (or reserved) quantities when no
quantities are set as done in the picking. As soon as a quantity is set,
it will revert to its previous behavior.

For that, it needs to use the `move_ids` instead of the `move_line_ids`
for the wizard, as it allows to check the initial demand while still
providing the `move_line_ids` when needed.

See odoo/odoo#106414

closes odoo/upgrade#4091

Signed-off-by: Christophe Simonis <chs@odoo.com>