[FIX] base/0.0.0: ensure valid view_mode
When removing dashboard view type from the window actions `view_mode` we
need to ensure they do not remain empty. In principle an empty value
means that the record was edited and thus we must enforce it is updated
during the upgrade. Allowing update is needed for 16.0 because we
removed dashboard view type. For the rest we just remove the dashboards
from the window actions if no dashboard view can be found.

```
...
   File "/home/odoo/src/odoo/16.0/odoo/addons/base/models/ir_ui_view.py", line 2616, in _get_view_cache
    arch, view = self._get_view(view_id, view_type, **options)
   File "/home/odoo/src/odoo/16.0/addons/web/models/models.py", line 241, in _get_view
    arch, view = super()._get_view(view_id, view_type, **options)
   File "/home/odoo/src/odoo/16.0/odoo/addons/base/models/ir_ui_view.py", line 2570, in _get_view
    raise UserError(_("No default view of type '%s' could be found !", view_type))
 odoo.exceptions.UserError: No default view of type '' could be found !
```

upg-420877

closes odoo/upgrade#4103

Signed-off-by: Christophe Simonis <chs@odoo.com>