[IMP] purchase_stock/saas~16.1: Allow multiple purchase lines per move
With the introduction of alternative Purchase Orders, having only a
single Purchase Order Line linked to a move wouldn't keep track of the
original moves that generated the Purchase Order once an alternative was
created for it.

In order to do that, it was necessary to transform the
`created_purchase_line_id` m2o into a `created_purchase_line_ids` m2m.

See odoo/odoo#106411

closes odoo/upgrade#4116

Signed-off-by: Christophe Simonis <chs@odoo.com>