[IMP] website/saas~16.2: move all pages popups to a dedicated area
Before 16.2, when the user wanted to have a popup displayed on every
page, we put this popup in the footer (because the footer is shared by
all pages). This dependency of the popup on the footer was a problem
(see community commit message), it caused bugs/limitations. So, from
16.2, the popups that must be visible on all pages are placed in a new
dedicated area. So this migration script just moves the existing popups
from the footer to the new area(`#o_shared_blocks`). For popups in
inactive footers, they are just removed.

See https://github.com/odoo/odoo/pull/107387

Related to opw-2971181

closes odoo/upgrade#4120

Related: odoo/odoo#107387
Signed-off-by: Christophe Simonis <chs@odoo.com>