[FIX] account/16.0.1.2: possible duplicate depending on order
To prevent temporary duplicates in the database
that would violate the unique constraint, we need
to ensure that we fix the account codes in the
correct order. To do this, we can add an ORDER BY
clause to specify the order in which the codes
should be fixed. This will ensure that the first
code we apply to the database is the one that
does not need to be changed.

upg-422567

closes odoo/upgrade#4121

Signed-off-by: Christophe Simonis <chs@odoo.com>