[IMP] base/0.0.0: re-activate children of disabled views
When we check views we disable its children first. If the view still
fails we disable it. At this point it's important to re-enable all
inheriting views that were disabled.

The benefits of re-enabling those views are:
1. We will correctly check the children views, that otherwise would be
   skipped since they are disabled.
2. We return a DB with _less_ disabled views. Thus it will be simpler
   for clients to restore and check them.

The fact that a view is disabled by setting `active = False` is relevant
if this view is referenced directly from an action. The ORM ignores the
active flag in such cases. Thus we need to ensure that when we return a
disabled view it was done correctly via the `disable` function that
correctly handles cleaning the actions.

closes odoo/upgrade#4129

Signed-off-by: Christophe Simonis <chs@odoo.com>
Co-authored-by: Alvaro Fuentes <afu@odoo.com>