[FIX] base/0.0.0: discard translated with null lang
When upgrading translations to the new jsonb system we get an error for
entries with `state=translated` and null `lang`.
```
            WITH t AS (
                SELECT res_id, jsonb_object_agg(lang, value) AS value
                  FROM _ir_translation
                 WHERE type = 'model' AND name = 'product.template,name' AND state = 'translated'
              GROUP BY res_id
            )
            UPDATE product_template m
               SET "name" =  t.value || m."name"
              FROM t
             WHERE t.res_id = m.id

ERROR: field name must not be null
```
It seems there is not valid flow (other than bugs or custom code) that
leads to this state. Moreover those terms are not taken into account for
translations anwyay since the language is missing.

upg-421998

closes odoo/upgrade#4164

Signed-off-by: Christophe Simonis <chs@odoo.com>